+++ web/satellite/src/components/project/ProjectNavigation.vue
@@ -1,0 +1,59 @@
+// Copyright (C) 2019 Storj Labs, Inc.
+// See LICENSE for copying information.
+
+<template>
+	<div class="project-navigation-container">
+		<router-link :to="navLink.path" class="project-navigation-container__item" v-for="navLink in navigation">
+			<p>{{navLink.name}}</p>
+		</router-link>
+	</div>
+</template>
+
+<script lang="ts">
+    import { Component, Vue } from 'vue-property-decorator';
+    import { PROJECT_ROUTES } from '@/utils/constants/projectNavigation';
+
+    @Component(
+        {
+            data: function() {
+                return {
+                    navigation: PROJECT_ROUTES,
+                };
+            },
+        }
+    )
+
+    export default class ProjectNavigation extends Vue {}
+</script>
+
+<style scoped lang="scss">
+    .project-navigation-container {
+	    width: auto;
+		display: flex;
+
+	    &__item {
+		    width: 150px;
+		    height: 48px;
+		    border-bottom: 1px solid #AFB7C1;
+		    display: flex;
+		    align-items: center;
+		    justify-content: center;
+
+		    p {
+				font-family: 'font_medium';
+			    font-size: 14px;
+			    line-height: 20px;
+			    color: #AFB7C1;
+		    }
+
+		    &.router-link-exact-active,
+		    &:hover {
+			    border-bottom: 3px solid #2683FF;
+
+			    p {
+				    color: #354049;
+			    }
+		    }
+	    }
+    }
+</style>
