+++ web/satellite/src/router/index.ts
@@ -9,11 +9,12 @@ import Register from '@/views/register/Register.vue';
 import ForgotPassword from '@/views/forgotPassword/ForgotPassword.vue';
 import Dashboard from '@/views/Dashboard.vue';
 import AccountArea from '@/components/account/AccountArea.vue';
-import ProjectDetailsArea from '@/components/project/ProjectDetailsArea.vue';
+import ProjectOverviewArea from '@/components/project/ProjectOverviewArea.vue';
 import TeamArea from '@/components/team/TeamArea.vue';
 import Page404 from '@/components/errors/Page404.vue';
 import ApiKeysArea from '@/components/apiKeys/ApiKeysArea.vue';
 import UsageReport from '@/components/project/UsageReport.vue';
+import ProjectDetails from '@/components/project/ProjectDetails.vue';
 import BucketArea from '@/components/buckets/BucketArea.vue';
 import { getToken } from '@/utils/tokenManager';
 import store from '@/store';
@@ -51,15 +52,32 @@ let router = new Router({
                     component: AccountArea
                 },
                 {
-                    path: ROUTES.PROJECT_DETAILS.path,
-                    name: ROUTES.PROJECT_DETAILS.name,
-                    component: ProjectDetailsArea
+                    path: ROUTES.PROJECT_OVERVIEW.path,
+                    name: ROUTES.PROJECT_OVERVIEW.name,
+                    component: ProjectOverviewArea,
+                    children: [
+                        {
+                            path: ROUTES.USAGE_REPORT.path,
+                            name: ROUTES.USAGE_REPORT.name,
+                            component: UsageReport,
+                        },
+                        {
+                            path: '',
+                            name: ROUTES.PROJECT_DETAILS.name,
+                            component: ProjectDetails
+                        },
+                        {
+                            path: ROUTES.PROJECT_DETAILS.path,
+                            name: ROUTES.PROJECT_DETAILS.name,
+                            component: ProjectDetails
+                        },
+                    ]
                 },
                 // Remove when dashboard will be created
                 {
                     path: '/',
                     name: 'default',
-                    component: ProjectDetailsArea
+                    component: ProjectOverviewArea
                 },
                 {
                     path: ROUTES.TEAM.path,
@@ -70,11 +88,6 @@ let router = new Router({
                     path: ROUTES.API_KEYS.path,
                     name: ROUTES.API_KEYS.name,
                     component: ApiKeysArea
-                },
-                {
-                    path: ROUTES.USAGE_REPORT.path,
-                    name: ROUTES.USAGE_REPORT.name,
-                    component: UsageReport,
                 },
                 // {
                 //     path: ROUTES.BUCKETS.path,
@@ -100,7 +113,7 @@ let router = new Router({
 // and if we are able to navigate to page without existing project
 router.beforeEach((to, from, next) => {
     if (isUnavailablePageWithoutProject(to.name as string)) {
-        next(ROUTES.PROJECT_DETAILS);
+        next(ROUTES.PROJECT_OVERVIEW);
 
         return;
     }
