+++ web/satellite/src/utils/constants/projectNavigation.ts
@@ -1,0 +1,18 @@
+export const PROJECT_ROUTES = {
+    DETAILS: {
+        path: '/project-overview/details',
+        name: 'Details'
+    },
+    PAYMENT_METHODS: {
+        path: '/project-overview/payment-methods',
+        name: 'Payment Methods'
+    },
+    BILLING_HISTORY: {
+        path: '/project-overview/billing-history',
+        name: 'Billing History'
+    },
+    REPORT: {
+        path: '/project-overview/usage-report',
+        name: 'Report'
+    },
+};
